+++ lib/ModelMixins/XmlRequestMixin.ts
@@ -15,15 +15,15 @@ export default function XmlRequestMixin<T extends AbstractConstructor<any>>(
       return loadXML(url);
     }
 
-    postXml(url: string, data: string) {
-      return loadWithXhr({
-        url: url,
-        method: "POST",
-        data,
-        overrideMimeType: "text/xml",
-        responseType: "document"
-      });
-    }
+    // postXml(url: string, data: string) {
+    //   return loadWithXhr({
+    //     url: url,
+    //     method: "POST",
+    //     data,
+    //     overrideMimeType: "text/xml",
+    //     responseType: "document"
+    //   });
+    // }
   }
 
   return XmlRequestMixin;