+++ lib/Models/ShareDataService.ts
@@ -59,16 +59,19 @@ export default class ShareDataService {
    * @return A promise for the token (which can later be resolved at /share/TOKEN).
    */
   async getShareToken(shareData: any): Promise<string> {
-    if (!this.isUsable) {
+    if (!this.isUsable || !this.url) {
       throw TerriaError.from("`ShareDataService` is not usable");
     }
 
     try {
       const result = await loadWithXhr({
-        url: this.url!,
+        url: this.url,
         method: "POST",
         data: JSON.stringify(shareData),
-        headers: { "Content-Type": "application/json" },
+        headers: {
+          "Content-Type": "application/json",
+          ...(await this.terria.configParameters.shareRequestHeaders?.())
+        },
         responseType: "json"
       });
       const json = typeof result === "string" ? JSON.parse(result) : result;
@@ -99,7 +102,16 @@ export default class ShareDataService {
       }
 
       return shareJson;
-    } catch (error) {
+    } catch (e: any) {
+      const error: Error & { statusCode?: number } = e || {};
+      if (error.statusCode === 404) {
+        throw TerriaError.from(error, {
+          title: i18next.t("models.shareData.shareNotFoundTitle"),
+          message: i18next.t("models.shareData.shareNotFoundMessage"),
+          importance: 1
+        });
+      }
+
       throw TerriaError.from(error, {
         title: i18next.t("models.shareData.expandErrorTitle"),
         message: i18next.t("models.shareData.expandErrorMessage", {