+++ lib/ReactViews/ExplorerWindow/Tabs/DataCatalogTab.tsx
@@ -1,5 +1,6 @@
 import { runInAction } from "mobx";
 import { observer } from "mobx-react";
+import { useCallback } from "react";
 import { useTranslation } from "react-i18next";
 import type CatalogMemberMixin from "../../../ModelMixins/CatalogMemberMixin";
 import Box from "../../../Styled/Box";
@@ -10,11 +11,18 @@ import Breadcrumbs from "../../Search/Breadcrumbs";
 import SearchBox, { DEBOUNCE_INTERVAL } from "../../Search/SearchBox";
 import Styles from "./data-catalog-tab.scss";
 import CatalogSearchProvider from "../../../Models/SearchProviders/CatalogSearchProvider";
+import MappableMixin from "../../../ModelMixins/MappableMixin";
 
 interface DataCatalogTabProps {
   items?: unknown[];
   searchPlaceholder?: string;
   onActionButtonClicked?: (item: CatalogMemberMixin.Instance) => void;
+  /** Override the default toggleItemOnMap behavior (when "add/remove from map" button is clicked in the data preview panel) */
+  onToggleItemOnMap?: (item: MappableMixin.Instance) => void;
+  hideToggleItemOnMap?: boolean;
+  hideSearch?: boolean;
+  hideActionButton?: boolean;
+  hideBreadcrumbs?: boolean;
 }
 
 const DataCatalogTab = observer(function DataCatalogTab(
@@ -26,7 +34,7 @@ const DataCatalogTab = observer(function DataCatalogTab(
   const {
     searchState,
     previewedItem: previewed,
-    breadcrumbsShown: showBreadcrumbs,
+    breadcrumbsShown: viewStateBreadcrumbsShown,
     terria
   } = viewState;
 
@@ -39,20 +47,20 @@ const DataCatalogTab = observer(function DataCatalogTab(
     });
   };
 
-  const search = () => {
+  const search = useCallback(() => {
     viewState.searchState.searchCatalog();
-  };
+  }, [viewState]);
 
   return (
     <div className={Styles.root}>
       <Box fullHeight column>
         <Box fullHeight overflow="hidden">
           <Box className={Styles.dataExplorer} styledWidth="40%">
-            {searchState.catalogSearchProvider && (
+            {!props.hideSearch && searchState.catalogSearchProvider && (
               <SearchBox
                 searchText={searchState.catalogSearchText}
                 onSearchTextChanged={changeSearchText}
-                onDoSearch={() => search()}
+                onDoSearch={search}
                 placeholder={searchPlaceholder}
                 debounceDuration={
                   terria.catalogReferencesLoaded
@@ -66,6 +74,7 @@ const DataCatalogTab = observer(function DataCatalogTab(
             <DataCatalog
               terria={terria}
               viewState={viewState}
+              hideActionButton={props.hideActionButton}
               onActionButtonClicked={props.onActionButtonClicked}
               items={props.items}
             />
@@ -75,11 +84,13 @@ const DataCatalogTab = observer(function DataCatalogTab(
               terria={terria}
               viewState={viewState}
               previewed={previewed}
+              onToggleItemOnMap={props.onToggleItemOnMap}
+              hideToggleItemOnMap={props.hideToggleItemOnMap}
             />
           </Box>
         </Box>
 
-        {showBreadcrumbs && (
+        {!props.hideBreadcrumbs && viewStateBreadcrumbsShown && (
           <Breadcrumbs
             terria={terria}
             viewState={viewState}