+++ lib/ReactViews/Mobile/MobileMenu.jsx
@@ -125,12 +125,19 @@ class MobileMenu extends Component {
             </div>
           ))}
           <div onClick={() => this.hideMenu()}>
-            <SettingPanel />
+            <SettingPanel
+              terria={this.props.terria}
+              viewState={this.props.viewState}
+              elementConfig={this.props.terria.elements.get(
+                "menu-bar-settings"
+              )}
+            />
           </div>
           <div onClick={() => this.hideMenu()}>
             <SharePanel
               terria={this.props.terria}
               viewState={this.props.viewState}
+              elementConfig={this.props.terria.elements.get("menu-bar-share")}
             />
           </div>
           {this.props.menuItems.map((menuItem) => (
@@ -141,20 +148,28 @@ class MobileMenu extends Component {
               {menuItem}
             </div>
           ))}
-          {mapUserGuide && <MobileMenuItem {...mapUserGuide} />}
+          {mapUserGuide && (
+            <MobileMenuItem>
+              <MobileMenuItem.Link {...mapUserGuide}>
+                {mapUserGuide.caption}
+              </MobileMenuItem.Link>
+            </MobileMenuItem>
+          )}
           {this.props.showFeedback && (
-            <MobileMenuItem
-              onClick={() => this.onFeedbackFormClick()}
-              caption={t("feedback.feedbackBtnText")}
-            />
+            <MobileMenuItem>
+              <MobileMenuItem.Button onClick={() => this.onFeedbackFormClick()}>
+                {t("feedback.feedbackBtnText")}
+              </MobileMenuItem.Button>
+            </MobileMenuItem>
           )}
           {hasStories && (
-            <MobileMenuItem
-              onClick={() => this.runStories()}
-              caption={t("story.mobileViewStory", {
-                storiesLength: this.props.terria.stories.length
-              })}
-            />
+            <MobileMenuItem>
+              <MobileMenuItem.Button onClick={() => this.runStories()}>
+                {t("story.mobileViewStory", {
+                  storiesLength: this.props.terria.stories.length
+                })}
+              </MobileMenuItem.Button>
+            </MobileMenuItem>
           )}
           {this.props.terria.configParameters.languageConfiguration
             ?.enabled && (
@@ -162,6 +177,7 @@ class MobileMenu extends Component {
               <LangPanel
                 terria={this.props.terria}
                 smallScreen={this.props.viewState.useSmallScreenInterface}
+                elementConfig={this.props.terria.elements.get("menu-bar-lang")}
               />
             </div>
           )}