+++ lib/ReactViews/StandardUserInterface/StandardUserInterface.tsx
@@ -66,6 +66,9 @@ const StandardUserInterfaceBase: FC<StandardUserInterfaceProps> = observer(
     });
 
     const handleDragOver = (e: DragEvent<HTMLDivElement>) => {
+      if (props.terria.configParameters.disableUserAddedData) {
+        return;
+      }
       if (!e.dataTransfer.types || !e.dataTransfer.types.includes("Files")) {
         return;
       }
@@ -76,6 +79,7 @@ const StandardUserInterfaceBase: FC<StandardUserInterfaceProps> = observer(
     };
 
     const shouldUseMobileInterface = () =>
+      !props.terria.configParameters.disableMobileInterface &&
       document.body.clientWidth < (props.minimumLargeScreenWidth ?? 768);
 
     const resizeListener = action(() => {
@@ -167,13 +171,15 @@ const StandardUserInterfaceBase: FC<StandardUserInterfaceProps> = observer(
               <div className={Styles.uiInner}>
                 {!props.viewState.hideMapUi && (
                   <>
-                    <Small>
-                      <MobileHeader
-                        menuItems={customElements.menu}
-                        menuLeftItems={customElements.menuLeft}
-                        version={props.version}
-                      />
-                    </Small>
+                    {!props.terria.configParameters.disableMobileInterface && (
+                      <Small>
+                        <MobileHeader
+                          menuItems={customElements.menu}
+                          menuLeftItems={customElements.menuLeft}
+                          version={props.version}
+                        />
+                      </Small>
+                    )}
                     <Medium>
                       <>
                         <WorkflowPanelPortal
@@ -222,7 +228,12 @@ const StandardUserInterfaceBase: FC<StandardUserInterfaceProps> = observer(
                   </div>
                 </Medium>
 
-                <section className={Styles.map}>
+                <section
+                  className={classNames(Styles.map, {
+                    [Styles.disableMobileInterface]:
+                      props.terria.configParameters.disableMobileInterface
+                  })}
+                >
                   <MapColumn
                     customElements={customElements}
                     animationDuration={animationDuration}