+++ test/Models/Catalog/esri/ArcGisMapServerCatalogItemSpec.ts
@@ -217,7 +217,7 @@ describe("ArcGisMapServerCatalogItem", function () {
       expect(item.layers).toBe("31");
     });
 
-    describe("when tokenUrl is set", function () {
+    xdescribe("when tokenUrl is set", function () {
       beforeEach(() => {
         runInAction(() => {
           item = new ArcGisMapServerCatalogItem(
@@ -287,7 +287,7 @@ describe("ArcGisMapServerCatalogItem", function () {
       });
     });
 
-    describe("when token is set", function () {
+    xdescribe("when token is set", function () {
       beforeEach(() => {
         runInAction(() => {
           item = new ArcGisMapServerCatalogItem(