+++ test/Models/ErrorServiceSpec.ts
@@ -2,7 +2,7 @@ import { http, HttpResponse } from "msw";
 import TerriaError from "../../lib/Core/TerriaError";
 import { ErrorServiceProvider } from "../../lib/Models/ErrorServiceProviders/ErrorService";
 import StubErrorServiceProvider from "../../lib/Models/ErrorServiceProviders/StubErrorServiceProvider";
-import Terria from "../../lib/Models/Terria";
+import Terria, { defaultLoadConfig } from "../../lib/Models/Terria";
 import { worker } from "../mocks/browser";
 
 describe("ErrorService", function () {
@@ -33,7 +33,7 @@ describe("ErrorService", function () {
   it("Initializes an error service, passing in config", async function () {
     const initSpy = spyOn(mockErrorServiceProvider, "init");
     await terria.start({
-      configUrl: "test-config.json",
+      loadConfig: () => defaultLoadConfig("test-config.json"),
       errorService: mockErrorServiceProvider
     });
     expect(initSpy).toHaveBeenCalledTimes(1);
@@ -42,7 +42,7 @@ describe("ErrorService", function () {
   it("Gets called with error", async function () {
     const errorSpy = spyOn(mockErrorServiceProvider, "error").and.callThrough();
     await terria.start({
-      configUrl: "test-config.json",
+      loadConfig: () => defaultLoadConfig("test-config.json"),
       errorService: mockErrorServiceProvider
     });
     const error = new TerriaError({
@@ -54,7 +54,7 @@ describe("ErrorService", function () {
 
   it("Falls back to stub provider", async () => {
     await terria.start({
-      configUrl: "test-config.json"
+      loadConfig: () => defaultLoadConfig("test-config.json")
     });
     expect(terria.errorService).toEqual(jasmine.any(StubErrorServiceProvider));
   });